refactor: adjust styling
