+++ src/components/Layout/CartCounter.module.scss
@@ -1,13 +1,13 @@
 .counter {
   position: absolute;
-  top: 28px;
-  right: 69px;
-  width: 1.2rem;
-  height: 1.2rem;
+  top: 32px;
+  right: 102px;
+  width: 1.5rem;
+  height: 1.5rem;
   border: 1px solid #fff;
   border-radius: 20px;
   text-align: center;
   font-size: 13px;
-  background-color: #f46805;
+  background-color: var(--orange-color);
   color: #fff;
 }
