+++ src/components/Layout/MenuCard.module.scss
@@ -1,13 +1,13 @@
 .card {
-  border-radius: 10px;
-  border-bottom-right-radius: 10px;
+  border-radius: 20px;
+  border-bottom-right-radius: 20px;
   background: #f5f5f5;
   box-shadow: 0 4px 16px rgba(23, 23, 24, 0.15);
 
   &-img {
     max-width: 100%;
-    border-top-left-radius: 2px;
-    border-top-right-radius: 10px;
+    border-top-left-radius: 20px;
+    border-top-right-radius: 20px;
     border-bottom-right-radius: 2px;
   }
 
@@ -18,13 +18,13 @@
     justify-content: space-between;
 
     & h3 {
-      font-size: 14px;
+      font-size: var(--normal-font-size);
     }
   }
 
   &-context {
     display: flex;
-    padding: 0.5rem 1rem 0.5rem;
+    padding: 0rem 1rem 0.5rem;
     justify-content: space-between;
     align-items: center;
 
@@ -33,26 +33,22 @@
       align-items: baseline;
     }
     & p {
-      font-size: 10px;
+      font-size: var(--small-font-size);
       font-weight: 600;
-      color: #f46805;
+      color: var(--orange-color);
     }
 
     & span {
       margin: 0 0.2rem;
-      font-size: 16px;
-      font-weight: 600;
+      font-size: var(--h3-font-size);
+      // font-weight: 600;
     }
 
     & button {
-      width: 3.5rem;
-      height: 1.8rem;
+      width: 6rem;
+      height: 2.2rem;
       border: none;
-      border-radius: 20px;
-      font-weight: 600;
-      background-color: #f46805;
-      color: #ffffff;
-      cursor: pointer;
+      font-size: var(--normal-font-size);
     }
   }
 }
