+++ src/navigation/Navigation.jsx
@@ -2,8 +2,8 @@ import React from "react";
 import { Route } from "react-router-dom";
 // import { BrowserRouter, Route } from "react-router-dom";
 // import Header from "../components/Layout/Header";
-import Home from "../components/Layout/Home/Home";
-import Menu from "../pages/Menu/index";
+import Home from "../pages/Home/Home";
+import Menu from "../pages/Menu/Menu";
 
 const Navigation = () => {
   return (
